refactor(governance): consolidate rules into 8 themed files (PR-FUNNEL-03)
Mechanical merge of 15 rules_*.go source files (+12 test files) into 8
themed files. No rule semantics, registration order, or public API
changes — verified by:

  - rule_inventory_test.go golden 81 IDs unchanged
  - gocell validate ./examples/{ssobff,todoorder,iotdevice} zero-diff
    (incl. --strict mode)
  - hack/verify-archtest-inventory.sh drift gate PASS

Final layout:

  rules_fmt.go              FMT-01..15, 24, 26..30, A1, REF-12 (relocated
                            from rules_ref.go — schemaRefs file existence
                            is I/O-flavored, pairs with FMT cluster)
  rules_ref.go              REF-01..11, 13..17 (REF-12 moved out)
  rules_topo.go             TOPO-01..09 (unchanged)
  rules_verify.go           VERIFY-01..06 (unchanged)
  rules_http.go             CH-04 + CH-05 + CH-06 (3 files merged;
                            parsedHandlerFile cache + findHandlerFile +
                            parseHandlerFile shared by CH-04/05 now in
                            one place)
  rules_misc_strict.go      strict orchestrator + FMT-16/17/19/C1/A1 +
                            FMT-20/21/22/23/25 (rules_strict.go +
                            rules_strict_extra.go merged; lineage-coupled
                            schema-walk helpers stay together)
  rules_misc_consistency.go CONTRACT-CONSISTENCY-EMIT-01 (renamed only)
  rules_misc_advisory.go    ADV-01..06 + OUTGUARD-01 + SLICE-CONSISTENCY-01
                            + FMT-19 impl + DOC-NAME-01 impl (5 files
                            merged; strict orchestrators in
                            rules_misc_strict.go cross-file-call the impls)

Tests follow source: rules_*_test.go themes mirror their source files
(rules_http_test.go / rules_misc_strict_test.go / rules_misc_advisory_test.go
new merges; rules_fmt_test.go / rules_topo_test.go / rules_misc_consistency_test.go
unchanged or pure renames).

Out-of-source-tree changes:

  - tools/archtest/httputil_invariants_test.go: hardcoded path
    rules_http_response_alignment.go → rules_http.go (3 occurrences)
  - docs/audit/archtest-inventory.md: regenerated by
    scripts/audit/list-archtests.sh (15→8 governance file count, rule
    membership re-grouped per new file structure)

Refs: docs/plans/202605070431-pr403-funnel-fix-roadmap.md PR-FUNNEL-03 §4
ref: kubernetes/staging/src/k8s.io/apimachinery/pkg/util/validation/
     (themed validation rule clusters by domain prefix)
