+++ kernel/governance/rules_fmt.go
@@ -1177,3 +1177,51 @@ func (v *Validator) validateFMT28() []ValidationResult {
 	}
 	return results
 }
+
+// =============================================================================
+// REF-12 — schema ref file existence (relocated from rules_ref.go in
+// PR-FUNNEL-03; the check is I/O-flavored and pairs with the FMT cluster's
+// disk-format rules rather than the REF cluster's metadata-graph rules).
+// =============================================================================
+
+// validateREF12 checks that contract.schemaRefs files exist on disk.
+// Skipped when root is empty.
+func (v *Validator) validateREF12() []ValidationResult {
+	if v.root == "" {
+		return nil
+	}
+	var results []ValidationResult
+	for _, c := range v.project.Contracts {
+		results = append(results, v.checkREF12Contract(c)...)
+	}
+	return results
+}
+
+// checkREF12Contract validates all schema refs declared by a single contract.
+func (v *Validator) checkREF12Contract(c *metadata.ContractMeta) []ValidationResult {
+	var results []ValidationResult
+	for _, ref := range metadata.ContractSchemaRefs(c) {
+		if ref.Ref == "" {
+			continue
+		}
+		resolved, err := metadata.ResolveContractSchemaRef(v.root, c, ref)
+		if err != nil {
+			results = append(results, v.newResult(
+				"REF-12", SeverityError, IssueInvalid,
+				contractFile(c),
+				ref.Field,
+				fmt.Sprintf("contract %q %s %q: %v", c.ID, ref.Field, ref.Ref, err),
+			))
+			continue
+		}
+		if !v.fileExists(resolved.AbsPath) {
+			results = append(results, v.newResult(
+				"REF-12", SeverityError, IssueRefNotFound,
+				contractFile(c),
+				ref.Field,
+				fmt.Sprintf("contract %q %s points to missing file %q", c.ID, ref.Field, ref.Ref),
+			))
+		}
+	}
+	return results
+}
