+++ kernel/governance/rules_misc_consistency.go
